+++ packages/orq-rc/src/models/operations/listdatasources.ts
@@ -224,7 +224,7 @@ export const ListDatasourcesData$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01KAJZB90WCEHR5STJEW5XPF91"),
+  _id: z.string().default("01KAK94K41KANZR7YCCFQ68HSE"),
   display_name: z.string(),
   description: z.string().optional(),
   status: ListDatasourcesStatus$inboundSchema,
@@ -267,7 +267,7 @@ export const ListDatasourcesData$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   ListDatasourcesData
 > = z.object({
-  id: z.string().default("01KAJZB90WCEHR5STJEW5XPF91"),
+  id: z.string().default("01KAK94K41KANZR7YCCFQ68HSE"),
   displayName: z.string(),
   description: z.string().optional(),
   status: ListDatasourcesStatus$outboundSchema,