+++ packages/orq-rc/RELEASES.md
@@ -5299,3 +5299,13 @@ Based on:
 - [typescript v4.0.0-rc.64] packages/orq-rc
 ### Releases
 - [NPM v4.0.0-rc.64] https://www.npmjs.com/package/@orq-ai/node/v/4.0.0-rc.64 - packages/orq-rc
+
+## 2025-11-21 13:21:38
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.659.0 (2.755.9) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.0.0-rc.65] packages/orq-rc
+### Releases
+- [NPM v4.0.0-rc.65] https://www.npmjs.com/package/@orq-ai/node/v/4.0.0-rc.65 - packages/orq-rc